pyproject.toml
[project]
name = "blink-waveforms-sabi"
version = "2.0.0"
description = "Demographics-adjusted evaluation of blink waveforms for sABI patients"
authors = [{ name = "Leonardo Corsi", email = "leonardo.corsi.app@gmail.com" }, { name = "Alfonso Magliacano", email = "amagliacano@dongnocchi.it" }]
readme = "README.md"
requires-python = ">=3.12.7"
license = { file = "LICENSE" }
dependencies = [
  "mne==1.8.0",
  "numpy==1.26.4",
  "pandas==2.2.2",
  "matplotlib==3.10.0",
  "seaborn==0.13.2",
  "scipy==1.13.1",
  "statsmodels==0.14.4",
  "scikit-posthocs==0.11.4",
  "scikit-learn==1.5.1",
  "joblib==1.4.2",
  "tqdm==4.66.5",
  "PyYAML==6.0.2",
  "python-dotenv==1.1.1",
  "ruptures==1.1.9",
  "statannotations==0.7.1",
  "ipython==8.27.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


[project.urls]
repository = "https://github.com/leonardocorsi/blink-waveforms-sabi"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E","F","I","UP","B","SIM"]
exclude = ["data", "results", ".venv"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
pythonpath = ["src","scripts"]
addopts = "-q"





